feature(cli): Development simpler workflow. Now generates 'Core' module in development.
This is tied to better and easier development of final product. We will generate by cli not only development, but prod too. This will ease out the updates and overall centralized updates to each project which will be generated by cli, because users should not change code in core by themselves to not to break it.